graph, store: Make HistoryEvent private to the Store
Since it does not appear in any public interfaces anymore, we can keep it
to just the store.
