+++ store/postgres/src/lib.rs
@@ -26,6 +26,7 @@ mod db_schema;
 mod entities;
 mod filter;
 mod functions;
+mod history_event;
 mod jsonb;
 mod notification_listener;
 mod relational;
